Move HandlerLoader implementations into the impl package.
