+++ mr/src/main/java/org/elasticsearch/hadoop/serialization/ScrollReaderConfigBuilder.java
@@ -23,7 +23,7 @@ import java.util.List;
 
 import org.elasticsearch.hadoop.cfg.Settings;
 import org.elasticsearch.hadoop.handler.HandlerLoader;
-import org.elasticsearch.hadoop.handler.PreloadedHandlerLoader;
+import org.elasticsearch.hadoop.handler.impl.PreloadedHandlerLoader;
 import org.elasticsearch.hadoop.serialization.builder.ValueReader;
 import org.elasticsearch.hadoop.serialization.dto.mapping.Mapping;
 import org.elasticsearch.hadoop.serialization.handler.read.DeserializationErrorHandler;
