+++ mr/src/main/java/org/elasticsearch/hadoop/serialization/handler/write/SerializationHandlerLoader.java
@@ -20,7 +20,7 @@
 package org.elasticsearch.hadoop.serialization.handler.write;
 
 import org.elasticsearch.hadoop.EsHadoopIllegalArgumentException;
-import org.elasticsearch.hadoop.handler.AbstractHandlerLoader;
+import org.elasticsearch.hadoop.handler.impl.AbstractHandlerLoader;
 
 public class SerializationHandlerLoader extends AbstractHandlerLoader<SerializationErrorHandler> {
